+++ lib/policy/policy.go
@@ -158,6 +158,15 @@ func ParseConfig(ctx context.Context, fin io.Reader, fname string, defaultDiffic
 			}
 		}
 
+		if b.RemoteAddressesURL != nil {
+			c, err := NewRemoteAddressesURLChecker(ctx, *b.RemoteAddressesURL, result.Logger.With("rule", b.Name))
+			if err != nil {
+				validationErrs = append(validationErrs, fmt.Errorf("while processing rule %s remote addresses url: %w", b.Name, err))
+			} else {
+				cl = append(cl, c)
+			}
+		}
+
 		if b.UserAgentRegex != nil {
 			c, err := NewUserAgentChecker(*b.UserAgentRegex)
 			if err != nil {